+++ src/graders/pass_graders.py
@@ -1,17 +1,10 @@
-from dataclasses import dataclass
-from typing import Mapping, Optional, Union
-
-from rlbot.training.training import Pass, Fail, Grade
+from rlbot.training.training import Pass, Grade
 from rlbottraining.grading.grader import Grader, TrainingTickPacket
-from rlbottraining.history.metric import Metric
+from typing import Optional, Union
+
+from .fail_graders import WrongGoalFail, FailOnTimeout
 
 
-class WrongGoalFail(Fail):
-    def __repr__(self):
-        return f'{super().__repr__()}: Ball went into the wrong goal.'
-
-
-@dataclass
 class PassOnGoalForAllyTeam(Grader):
     """
     Terminates the Exercise when any goal is scored.
@@ -19,8 +12,12 @@ class PassOnGoalForAllyTeam(Grader):
     otherwise returns a Fail.
     """
 
-    ally_team: int  # The team ID, as in game_datastruct.PlayerInfo.team
-    init_score: Optional[Mapping[int, int]] = None  # team_id -> score
+    def __init__(self, ally_team: int):
+        """
+        :param ally_team: number equal to game_datastruct.PlayerInfo.team.
+        """
+        self.ally_team = ally_team
+        self.init_score = None
 
     def on_tick(self, tick: TrainingTickPacket) -> Optional[Union[Pass, WrongGoalFail]]:
         score = {
@@ -74,43 +71,6 @@ class PassOnBallGoingAwayFromGoal(Grader):
 
         if self.consequtive_good_ticks >= self.REQUIRED_CONSECUTIVE_TICKS:
             return Pass()
-
-
-class FailOnTimeout(Grader):
-    """Fails the exercise if we take too long."""
-
-    class FailDueToTimeout(Fail):
-        def __init__(self, max_duration_seconds):
-            self.max_duration_seconds = max_duration_seconds
-
-        def __repr__(self):
-            return f'{super().__repr__()}: Timeout: Took longer than {self.max_duration_seconds} seconds.'
-
-    def __init__(self, max_duration_seconds: float):
-        self.max_duration_seconds = max_duration_seconds
-        self.initial_seconds_elapsed: float = None
-        self.measured_duration_seconds: float = None
-
-    def on_tick(self, tick: TrainingTickPacket) -> Optional[Grade]:
-        seconds_elapsed = tick.game_tick_packet.game_info.seconds_elapsed
-        if self.initial_seconds_elapsed is None:
-            self.initial_seconds_elapsed = seconds_elapsed
-        self.measured_duration_seconds = seconds_elapsed - self.initial_seconds_elapsed
-        if self.measured_duration_seconds > self.max_duration_seconds:
-            return self.FailDueToTimeout(self.max_duration_seconds)
-
-    @dataclass(frozen=True)
-    class TimeoutMetric(Metric):
-        max_duration_seconds: float
-        initial_seconds_elapsed: float
-        measured_duration_seconds: float
-
-    def get_metric(self) -> Optional[Metric]:
-        return FailOnTimeout.TimeoutMetric(
-            self.max_duration_seconds,
-            self.initial_seconds_elapsed,
-            self.measured_duration_seconds,
-        )
 
 
 class PassOnTimeout(FailOnTimeout):
